+++ model/db/attribution.gen.go
@@ -8,9 +8,9 @@ const TableNameAttribution = "Attribution"
 
 // Attribution mapped from table <Attribution>
 type Attribution struct {
-	ID         int32 `gorm:"column:id;primaryKey" json:"-"`
-	ContextID  int32 `gorm:"column:context_id;not null" json:"-"`
-	ArtifactID int32 `gorm:"column:artifact_id;not null" json:"-"`
+	ID         int64 `gorm:"column:id;primaryKey" json:"-"`
+	ContextID  int64 `gorm:"column:context_id;not null" json:"-"`
+	ArtifactID int64 `gorm:"column:artifact_id;not null" json:"-"`
 }
 
 // TableName Attribution's table name
