+++ model/db/execution.gen.go
@@ -8,8 +8,8 @@ const TableNameExecution = "Execution"
 
 // Execution mapped from table <Execution>
 type Execution struct {
-	ID                       int32  `gorm:"column:id;primaryKey" json:"-"`
-	TypeID                   int32  `gorm:"column:type_id;not null" json:"-"`
+	ID                       int64  `gorm:"column:id;primaryKey" json:"-"`
+	TypeID                   int64  `gorm:"column:type_id;not null" json:"-"`
 	LastKnownState           int32  `gorm:"column:last_known_state;not null" json:"-"`
 	Name                     string `gorm:"column:name;not null" json:"-"`
 	ExternalID               string `gorm:"column:external_id;not null" json:"-"`
